Fix HTML tables #68
Added table injection logic to a new JS file called hugo-custom.js,
created by an updated webpack file. The logic was added to the footer
after solstice injection (to ensure Bootstrap is loaded) in a new
footer_hugo partial. Variables for the table injection are set within
the head to be used in the page.

Change-Id: Ib99d75ded1caaf58241771bcefa176543016d68f
Signed-off-by: Martin Lowe <martin.lowe@eclipse-foundation.org>